files/requirements.txt
# ─── Core requirements ────────────────────────────────────────────────────────
ultralytics>=8.0.0          # YOLOv8 (auto-downloads model weights from COCO)
opencv-python>=4.8.0        # Camera capture & image drawing
numpy>=1.24.0               # Array processing
PyYAML>=6.0                 # Dataset config files
Pillow>=9.5.0               # Image utilities
matplotlib>=3.7.0           # Training plots (used by ultralytics)
seaborn>=0.12.0             # Confusion matrix plots

# ─── Optional: dataset download ───────────────────────────────────────────────
roboflow>=1.1.0             # Roboflow dataset API (free tier)

# ─── Optional: GPU acceleration ───────────────────────────────────────────────
# torch>=2.0.0              # Auto-installed by ultralytics
# torchvision>=0.15.0
